Updated Character Subtraction
Found a bug that referenced memory locations outside the the bounds of the memory allocated for Cstrings (standard vectors for the win again).
This updated code fixes the print character that called a while loop that returned false when it hits the null character, but did not work as intended. Therefore I modified the code to simply stop iterating when the index is greater than or equal to the size of the array (using the strlen() function). 

Hopefully this reduces weird output but no promises.